Hook up personality value to hue shifting code!!
Works at the moment by hijacking any call to GetPkmnData or GetBoxPkmnData to
write the personality value to my "free RAM". This is then loaded by the
hue shifting hijack. The assumption is that the last call to either function
prior to the hue shifting hijack should be for the same Pokemon.

Some code was added to clear the free RAM when the Pokedex is opened.